Added table merge util
