+++ Source/Internal/Table/PDFTableObject.swift
@@ -37,7 +37,7 @@ internal class PDFTableObject: PDFRenderObject {
     override internal func calculate(generator: PDFGenerator, container: PDFContainer) throws -> [(PDFContainer, PDFRenderObject)] {
         try PDFTableValidator.validateTable(table: table)
 
-        calculateMergedTable()
+        let mergedCells = PDFTableMergeUtil.calculateMerged(table: table)
 
         var cellItems: [[(cell: PDFTableCell, style: PDFTableCellStyle, frames: (cell: CGRect, content: CGRect))]] = []
 
@@ -139,49 +139,6 @@ internal class PDFTableObject: PDFRenderObject {
         try PDFOffsetObject(offset: finalOffset).calculate(generator: generator, container: container)
 
         return renderObjects.objects
-    }
-
-    internal func calculateMergedTable() -> [[PDFTableNode]] {
-        guard table.size.rows > 0 && table.size.columns > 0 else {
-            return []
-        }
-        var found = table.cells.map({ $0.map({ _ in false }) })
-
-        var rows: [[PDFTableNode]] = []
-        // Iterate all rows
-        for (rowIdx, row) in table.cells.enumerated() {
-            var rowNodes: [PDFTableNode] = []
-
-            for (colIdx, cell) in row.enumerated() where !found[rowIdx][colIdx] {
-                let node = PDFTableNode(cell: cell, position: PDFTableCellPosition(row: rowIdx, column: colIdx))
-                found[rowIdx][colIdx] = true
-                for adjacentColumnIdx in (colIdx + 1)..<table.size.columns {
-                    guard row[adjacentColumnIdx] === cell else {
-                        break
-                    }
-                    found[rowIdx][adjacentColumnIdx] = true
-                    node.moreColumnsSpan += 1
-                }
-                for adjacentRowIdx in (rowIdx + 1)..<table.size.rows {
-                    // Cell underneath current cell in next row
-                    guard table.cells[adjacentRowIdx][colIdx] === cell else {
-                        break
-                    }
-                    // Check if all statisfy identical
-                    guard table.cells[adjacentRowIdx][0...node.moreColumnsSpan].allSatisfy({ $0 === cell }) else {
-                        break
-                    }
-                    node.moreRowsSpan += 1
-                    for columnIdxInAdjacentRow in colIdx..<(colIdx + node.moreColumnsSpan) {
-                        found[adjacentRowIdx][columnIdxInAdjacentRow] = true
-                    }
-                }
-                rowNodes.append(node)
-            }
-
-            rows.append(rowNodes)
-        }
-        return rows
     }
 
     /**
